// src/app/page/master/map-keteranganalasan-status/map-keteranganalasan-status.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapKeteranganalasanStatus {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	value?: any;
	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}